requirements.txt
streamlit==1.28.0
ansible-runner==2.3.4
ansible==8.5.0
pandas==2.1.1
pyyaml==6.0.1
paramiko==3.3.1
cryptography==41.0.5
python-dotenv==1.0.0